+++ recommendation-service/src/main/java/com/infogain/repository/RecommendRepository.java
@@ -12,13 +12,13 @@ import com.infogain.entity.RecommendEntity;
 @Repository
 public interface RecommendRepository extends JpaRepository<RecommendEntity, Serializable> {
 
-	RecommendEntity findByUserIdAndMovieId(Integer userId, Integer movieId);
+	RecommendEntity findByUserIdAndMovieId(Long userId, Long movieId);
 
-	List<RecommendEntity> findByUserIdInAndIsLiked(List<Integer> userIds, boolean isLiked);
+	List<RecommendEntity> findByUserIdInAndIsLiked(List<Long> userIds, boolean isLiked);
 
 	@Query(value = "SELECT TOP 5 movie_id as movieId, count(*) as likes FROM recommend_metadata WHERE movie_id IN ?1 "
 			+ "GROUP BY movie_id HAVING is_movie_liked=true ORDER BY count(*) DESC", nativeQuery = true)
-	List<MovieDTO> getRecommendedTop5MovieIds(List<Integer> movieIds);
+	List<MovieDTO> getRecommendedTop5MovieIds(List<Long> movieIds);
 	
 	@Query(value = "SELECT TOP 5 movie_id as movieId, count(*) as likes FROM recommend_metadata "
 			+ "GROUP BY movie_id HAVING is_movie_liked=true ORDER BY count(*) DESC", nativeQuery = true)
