+++ recommendation-service/src/main/java/com/infogain/request/LikeRequest.java
@@ -5,19 +5,19 @@ import javax.validation.constraints.NotNull;
 public class LikeRequest {
 
 	@NotNull(message = "movieId can not be null or empty")
-	private Integer movieId;
+	private Long movieId;
 	@NotNull(message = "userId can not be null or empty")
-	private Integer userId;
-	public Integer getMovieId() {
+	private Long userId;
+	public Long getMovieId() {
 		return movieId;
 	}
-	public void setMovieId(Integer movieId) {
+	public void setMovieId(Long movieId) {
 		this.movieId = movieId;
 	}
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
